ml/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
ultralytics
python-multipart